gnomeos: We can now build gobject-introspection

+++ src/ostbuild/pyostbuild/builtin_compile_one.py
@@ -36,42 +36,11 @@ _BLACKLIST_REGEXPS = map(re.compile,
 _DEVEL_REGEXPS = map(re.compile,
                      [r'/usr/include/',
                       r'/usr/share/pkgconfig/',
+                      r'/usr/share/aclocal/',
                       r'/(?:usr/)lib(?:|(?:32)|(?:64))/pkgconfig/.*\.pc$',
                       r'/(?:usr/)lib(?:|(?:32)|(?:64))/[^/]+\.so$'
+                      r'/(?:usr/)lib(?:|(?:32)|(?:64))/[^/]+\.a$'
                       ])
-
-class BuildSystemScanner(object):
-    @classmethod
-    def _find_file(cls, names):
-        for name in names:
-            if os.path.exists(name):
-                return name
-        return None
-
-    @classmethod
-    def get_configure_source_script(cls):
-        return cls._find_file(('./configure.ac', './configure.in'))
-
-    @classmethod
-    def get_configure_script(cls):
-        return cls._find_file(('./configure', ))
-
-    @classmethod
-    def get_bootstrap_script(cls):
-        return cls._find_file(('./autogen.sh', ))
-
-    @classmethod
-    def get_silent_rules(cls):
-        src = cls.get_configure_source_script()
-        if not src:
-            return False
-        f = open(src)
-        for line in f:
-            if line.find('AM_SILENT_RULES') >= 0:
-                f.close()
-                return True
-        f.close()
-        return False
 
 class OstbuildCompileOne(builtins.Builtin):
     name = "compile-one"
@@ -81,25 +50,12 @@ class OstbuildCompileOne(builtins.Builtin):
         builtins.Builtin.__init__(self)
         self.tempfiles = []
 
-    def _search_file(self, filename, pattern):
-        f = open(filename)
-        for line in f:
-            if line.startswith(pattern):
-                f.close()
-                return line
-        f.close()
-        return None
-
-    def _find_buildapi_makevariable(self, name, builddir='.'):
-        var = '.%s:' % (name, )
-        line = None
-        path = os.path.join(builddir, 'Makefile.in')
-        if os.path.exists(path):
-            line = self._search_file(path, var)
-        path = os.path.join(builddir, 'Makefile')
-        if not line and os.path.exists(path):
-            line = self._search_file(path, var)
-        return line is not None
+    def _has_buildapi_configure_variable(self, name):
+        var = '#buildapi-variable-%s' % (name, )
+        for line in open('configure'):
+            if line.find(var) >= 0:
+                return True
+        return False
 
     def execute(self, args):
         self.default_buildapi_jobs = ['-j', '%d' % (cpu_count() * 2, )]
@@ -134,9 +90,9 @@ class OstbuildCompileOne(builtins.Builtin):
 
         for arg in args:
             if arg.startswith('--ostbuild-resultdir='):
-                self.ostbuild_resultdir=arg[len('ostbuild-resultdir='):]
-            elif arg.startswith('ostbuild-meta='):
-                self.ostbuild_meta=arg[len('ostbuild-meta='):]
+                self.ostbuild_resultdir=arg[len('--ostbuild-resultdir='):]
+            elif arg.startswith('--ostbuild-meta='):
+                self.ostbuild_meta=arg[len('--ostbuild-meta='):]
             elif arg.startswith('--'):
                 self.configargs.append(arg)
             else:
@@ -145,10 +101,10 @@ class OstbuildCompileOne(builtins.Builtin):
         self.metadata = {}
 
         if self.ostbuild_meta is None:
-            output = subprocess.check_output(['ostbuild-autodiscover-meta'])
+            output = subprocess.check_output(['ostbuild', 'autodiscover-meta'])
             ostbuild_meta_f = StringIO(output)
         else:
-            ostbuild_meta_f = open(ostbuild_meta)
+            ostbuild_meta_f = open(self.ostbuild_meta)
 
         for line in ostbuild_meta_f:
             (k,v) = line.split('=', 1)
@@ -160,46 +116,33 @@ class OstbuildCompileOne(builtins.Builtin):
             if k not in self.metadata:
                 fatal('Missing required key "%s" in metadata' % (k, ))
 
-        self.phase_bootstrap()
+        autogen_script = None
+        if not os.path.exists('configure'):
+            log("No 'configure' script found, looking for autogen/bootstrap")
+            for name in ['autogen', 'autogen.sh', 'bootstrap']:
+                if os.path.exists(name):
+                    log("Using bootstrap script '%s'" % (name, ))
+                    autogen_script = name
+            if autogen_script is None:
+                fatal("No configure or autogen script detected; unknown buildsystem")
 
-    def phase_bootstrap(self):
-        have_configure = BuildSystemScanner.get_configure_script() 
-        have_configure_source = BuildSystemScanner.get_configure_source_script()
-        if not (have_configure or have_configure_source):
-            fatal("No configure or bootstrap script detected; unknown buildsystem")
-            return
-    
-        need_v1 = BuildSystemScanner.get_silent_rules()
-        if need_v1:
-            log("Detected AM_SILENT_RULES, adding --disable-silent-rules to configure")
-            self.configargs.append('--disable-silent-rules')
-    
-        if have_configure:
-            self.phase_configure()
+        if autogen_script is not None:
+            env = dict(os.environ)
+            env['NOCONFIGURE'] = '1'
+            run_sync(['./' + autogen_script], env=env)
         else:
-            bootstrap = BuildSystemScanner.get_bootstrap_script()
-            if bootstrap:
-                log("Detected bootstrap script: %s, using it" % (bootstrap, ))
-                args = [bootstrap]
-                # Add NOCONFIGURE; GNOME style scripts use this
-                env = dict(os.environ)
-                env['NOCONFIGURE'] = '1'
-                run_sync(args, env=env)
-            else:
-                log("No bootstrap script found; using generic autoreconf")
-                run_sync(['autoreconf', '-f', '-i'])
-            self.phase_configure()
+            log("Using existing 'configure' script")
     
-    def phase_configure(self):
         use_builddir = True
-        doesnot_support_builddir = self._find_buildapi_makevariable('buildapi-no-builddir')
+        doesnot_support_builddir = self._has_buildapi_configure_variable('no-builddir')
         if doesnot_support_builddir:
-            log("Found .buildapi-no-builddir; copying source tree to _build")
-            shutil.rmtree('_build')
-            os.mkdir('_build')
+            log("Found no-builddir Build API variable; copying source tree to _build")
+            if os.path.isdir('_build'):
+                shutil.rmtree('_build')
             shutil.copytree('.', '_build', symlinks=True,
                             ignore=shutil.ignore_patterns('_build'))
             use_builddir = False
+            builddir = '.'
     
         if use_builddir:
             builddir = '_build'
@@ -207,26 +150,20 @@ class OstbuildCompileOne(builtins.Builtin):
             if not os.path.isdir(builddir):
                 os.mkdir(builddir)
     
-        configstatus = 'config.status'
-        if not os.path.exists(configstatus):
-            if use_builddir:
-                args = ['../configure']
-            else:
-                args = ['./configure']
-            args.extend(self.configargs)
-            if use_builddir:
-                run_sync(args, cwd=builddir)
-            else:
-                run_sync(args)
+        if use_builddir:
+            args = ['../configure']
         else:
-            log("Found %s, skipping configure" % (configstatus, ))
-        self.phase_build(builddir=builddir)
-    
-    build_status = False
-    
-    def phase_build(self, builddir=None):
-        if not os.path.exists(os.path.join(builddir, 'Makefile')):
+            args = ['./configure']
+        args.extend(self.configargs)
+        if use_builddir:
+            run_sync(args, cwd=builddir)
+        else:
+            run_sync(args)
+
+        makefile_path = os.path.join(builddir, 'Makefile')
+        if not os.path.exists(makefile_path):
             fatal("No Makefile found")
+
         args = list(self.makeargs)
         user_specified_jobs = False
         for arg in args:
@@ -234,35 +171,18 @@ class OstbuildCompileOne(builtins.Builtin):
                 user_specified_jobs = True
     
         if not user_specified_jobs:
-            notparallel = self._find_buildapi_makevariable('NOTPARALLEL', builddir=builddir)
-            if not notparallel:
+            has_notparallel = False
+            for line in open(makefile_path):
+                if line.startswith('.NOTPARALLEL'):
+                    has_notparallel = True
+                    log("Found .NOTPARALLEL")
+
+            if not has_notparallel:
                 log("Didn't find NOTPARALLEL, using parallel make by default")
                 args.extend(self.default_buildapi_jobs)
     
         run_sync(args, cwd=builddir)
-    
-        self.phase_make_artifacts(builddir=builddir)
-    
-    def make_artifact(self, name, from_files, tempdir=None, resultdir=None):
-        targz_name = name + '.tar.gz'
-        (fd,filelist_temp)=tempfile.mkstemp(prefix='ostree-filelist-%s' % (name, ))
-        os.close(fd)
-        self.tempfiles.append(filelist_temp)
-        f = open(filelist_temp, 'w')
-        for filename in from_files:
-            assert ('\n' not in filename)
-            f.write(filename)
-            f.write('\n')
-        f.close()
-        if resultdir:
-            result_path = os.path.join(resultdir, targz_name)
-        else:
-            result_path = targz_name
-        args = ['tar', '-c', '-z', '-C', tempdir, '-f', result_path, '-T', filelist_temp]
-        run_sync(args)
-        log("created: %s" % (os.path.abspath (result_path), ))
-    
-    def phase_make_artifacts(self, builddir=None):
+
         name = self.metadata['NAME']
         assert ',' not in name
         branch = self.metadata['BRANCH']
@@ -280,7 +200,7 @@ class OstbuildCompileOne(builtins.Builtin):
     
         artifact_prefix='artifact-%s,%s,%s,%s,%s' % (root_name, root_version, name, branch, version)
 
-        tempdir = tempfile.mkdtemp(prefix='ostree-build-%s-' % (name,))
+        tempdir = tempfile.mkdtemp(prefix='ostbuild-%s-' % (name,))
         self.tempfiles.append(tempdir)
         args = ['make', 'install', 'DESTDIR=' + tempdir]
         run_sync(args, cwd=builddir)
@@ -318,9 +238,6 @@ class OstbuildCompileOne(builtins.Builtin):
             self.make_artifact(artifact_prefix + '-devel', devel_files, tempdir=tempdir, resultdir=self.ostbuild_resultdir)
         self.make_artifact(artifact_prefix + '-runtime', runtime_files, tempdir=tempdir, resultdir=self.ostbuild_resultdir)
 
-        self.phase_complete()
-
-    def phase_complete(self):
         for tmpname in self.tempfiles:
             assert os.path.isabs(tmpname)
             if os.path.isdir(tmpname):
@@ -331,5 +248,24 @@ class OstbuildCompileOne(builtins.Builtin):
                     pass
                 except OSError, e:
                     pass
-
+    
+    def make_artifact(self, name, from_files, tempdir=None, resultdir=None):
+        targz_name = name + '.tar.gz'
+        (fd,filelist_temp)=tempfile.mkstemp(prefix='ostbuild-filelist-%s' % (name, ))
+        os.close(fd)
+        self.tempfiles.append(filelist_temp)
+        f = open(filelist_temp, 'w')
+        for filename in from_files:
+            assert ('\n' not in filename)
+            f.write(filename)
+            f.write('\n')
+        f.close()
+        if resultdir:
+            result_path = os.path.join(resultdir, targz_name)
+        else:
+            result_path = targz_name
+        args = ['tar', '-c', '-z', '-C', tempdir, '-f', result_path, '-T', filelist_temp]
+        run_sync(args)
+        log("created: %s" % (os.path.abspath (result_path), ))
+    
 builtins.register(OstbuildCompileOne)
